Break out logic for getting a user's notification preferences to a separate function
